Refactored the Hash class.
Added a tableSize member to simplify functions that need the size
of the table but used to call the table's member getCount().

Added a few member functions to improve readability and remove
code duplication.

Changed the type of the count member to size_t.
Most of the functions that take an integer as argument had an int
parameter, which is now of type size_t.