Add two new methods to the LinkedList class
The only way to access the first or the last element in a list was
through an iterator object which was inconvenient.

Now, with the inclusion of the two new methods getFirst and getLast,
the first and the last element can be accessed directly too. The
methods, however, return a const reference to an element. So, to be
modified, the element must be accessed through an iterator object.