Add abstract methods to the iterator classes
Instances of the iterator classes for LinkedList and DynamicArray
can not be returned by classes that use them in their representation
without breaking encapsulation. Further more, it is not flexible as
client code would have to change, if that representation changes.

That's why an abstract Iterator class was introduced and LinkedList-
Iterator and DynamicArrayIterator are modified to be its subclasses.