Change Graph's interface
Add a public getVertexWithIdentifier method to the interface.

Most of the methods in the current version take vertex identifiers as
arguments and have to search for the vertices before doing anything
else.

The methods mentioned above can now take the vertices themselves as
arguments, which is more efficient since the client code already has
access to them and there is no need of an extra search.