Replace exceptions with assertions in DynamicArray
Some methods that take an index as an argument throw an exception if
the index is invalid. Since this index does not come as an input to the
program, if it is invalid, this means that there is an error in the
source code and this is not an exceptional event to handle.