Change the return type of a Graph's method
The current version of getVertexWithIdentifier returns a pointer to a
vertex or a null pointer (if no vertex has a matching identifier).
This enforces each caller to check the returned pointer.

The new version returns an lvalue reference to a vertex. If no vertex
has a matching identifier, an exception is thrown since the input was
invalid.