Remove DynamicArray's search method
It is not used in the application and the reason probably is that it
uses equality test on the items themselves and not on a specified key
of the items. Moreover, the method adds a type requirement for the Item
class, that it must have overloaded operator==.
The only good reason to keep it might be for a SortedDynamicArray
implementation that implements the method using binary search. But it
doesn't looke like such an implementation will be needed in the near
future.