Add a KeyAccessor template parameter to PriorityQueue
The algorithms that will use a priority queue will need the keys of
items that are inserted in the queue to be attributes of the items.
This means that the PriorityQueue class need not store a pair of an
item and a key, but just the item itself. To be able to access the key
of on item, PriorityQueue needs a functor (a KeyAccessor).