Modify the ShortestPathAlgorithm base class
The responsibility for registering an object of type ShortestPathAlgo-
rithm is not the class', the constructor now does not take care of
that.

There shouldn't be two algorithms with the same identifier. But this
can be ensured by the client code. There is no need to delete the copy
constructor and assignment, the class can now use the compiler supplied
ones.

Subclasses use initialiseSource method and their implementations have
common statements, move the method to the base class and provide these
statements as a default implementation.