Refactor the BFS class
Remove the 'theOnlyInstance' static variable and make the constructor, copy-constructor
and copy-assignment public. Only one instance of the class will be used by the application
but this need not be enforced by the class. It should be independent.

Replace the name 'destination' with 'target' as it seems more appriopriate.

Reuse the base class' implementation of the initialiseSource method.