Refactor the IterativeDeepeningDFS class
Remove the static variable 'theOnlyInstance'. Its purpose is to re-
gister the only needed instance via the base class. The base class no
longer registers instances as it's not its responsibility.

Make the constructor, copy-constructor and copy-assignment public. They
are private to enforce having only one instance of the class. This
need not be enforced by the class itself. It should be independent, the
code that sets up the system will make sure to instantiate the class
just once.

Replace the name 'destination' with 'target' to be consistent with
naming in the whole hierarchy.