Refactor the Edge class
Replace an exception with an assertion in setVertex.
If the argument is a null pointer, that is not an exceptional case,
it is an error in the source code.

Make the destructor and copy-constructor public to be able to
remove an odd friend declaration.

Add the missing special members declarations.