Make the calls to isValid explicit in LinkedList
The methods of LinkedList call LinkedListIterator::operator bool() to
check if the iterator object is valid. Other parts of the system call
the isValid method. It isn't consistent calling different methods for
the same thing and the isValid call is easier to read, so the calls in
LinkedList will change.