Make two private methods protected in GraphBase
The two methods that return concrete iterators for vertices and edges
are private to keep the class loosely coupled with its subclasses.

Since now the public iterator methods return const iterators and the
subclasses may need methods that provide non-const iterators, the two
private methods will be made protected. This way the subclasses will be
a little more coupled with their base but the typedef declarations
still make it easy to change the base class' representation without
affecting subclasses.

The other solution would be to add two new methods that return abstract
iterators but this would make it too many methods just for iterators
and iteration would be less efficient.