Fix #1
When after a removal the items occupy a very small percentage
of the slots in the table, the class tries to shrink it so
that space is not wasted.

In case there is not enough memory for a new (smaller) table, the
cluster following the now empty slot is rehashed.