Modify PriorityQueue's move-assignment operator
All the special members of the class are consistent when handles are
not used. For example when the elements are just integers.

When (pointers to) external to the queue objects are stored and handles
are used, the copy semantics are not to be used. The reason is that if
one object is copied, two objects update the handles of the same set of
external elements. Consequently, if the key of one of the elements is
decreased in one of the queues, its handle will be modified by that
queue and that handle will no longer be consistent with the other
queue.