Pull up common code from ShortestPathAlgorithm subclasses
All the subclasses implement the findShortestPath method in a similar
way. They first initialise some state, execute a specific algorithm,
create a Path instance, clean up the state used in the call and return
the path.

This is why ShortestPathAlgorithm::findShortestPath method can use the
Template Method pattern and get rid of the duplicate code and the
reoccurence of the same idea.