Add a removeLast method to DynamicArray
A lot of client code does this operation on arrays but needs to obtain
the items' count first and then call removeAt. Add this new method to
simplify client code.

Also increment the count member in the add method after the assignment
completed, this way consistensy is guaranteed if the assignment fails.