Prioritise move-assignment in DynamicArray shift methods
If move-assignment is available, it is very likely that even if it is
potentially throwable, it still is less likely to throw than copy-
assignment. And if an exception occurs in the middle of shifting array
elements, changes can not be undone no matter which of the two
assignment operators is used.