Change DirectoryLoader's algorithm
If an exception is raised (most probably because of an invalid input
file), the current version logs the error message and continues with
the rest of the files in the directory.

Modify the algorithm so that a potential exception terminates it, that
is, graphs loaded so far are destroyed as the exception propagates to
the caller.