Change the access specifier for a Command method
Commands with no parameters still have to parse their arguments (to see
if any were passed for example).Since this isn't intuitive, the Command
class offers it as a default behaviour for the parseArguments method.
It is declared as a pure-virtual method so that subclasses explicitly
ask for this default behaviour and not fall back on it unintentionally.
That's why the method has to become protected.